Feature/gui updates
* fixing outdated cv2 reverences

* moved default parameter settings to a yaml file, merged MARS and MARS_v1_8.py

* refactored MARS and MARS_gui

Created common library MARS_queue that can be called by either to run MARS
