+++ MARS_environment_linux.yml
@@ -2,132 +2,135 @@ name: mars
 channels:
   - conda-forge
   - anaconda
-  - intel
   - defaults
 dependencies:
-  - _py-xgboost-mutex=2.0=cpu_0
-  - ca-certificates=2020.7.22=0
-  - certifi=2020.6.20=py36_0
-  - cloudpickle=1.6.0=py_0
-  - cycler=0.10.0=py36_0
-  - cytoolz=0.11.0=py36h7b6447c_0
-  - dask-core=2.27.0=py_0
-  - dbus=1.13.12=h746ee38_0
-  - decorator=4.4.2=py_0
-  - dill=0.3.2=py_0
-  - expat=2.2.9=he6710b0_2
-  - fontconfig=2.13.0=h9420a91_0
-  - freetype=2.10.2=h5ab3b9f_0
-  - glib=2.56.2=hd408876_0
-  - gst-plugins-base=1.14.0=hbbd80ab_1
-  - gstreamer=1.14.0=hb453b48_1
-  - icu=58.2=he6710b0_3
-  - imageio=2.9.0=py_0
-  - joblib=0.16.0=py_0
-  - jpeg=9b=habf39ab_1
-  - kiwisolver=1.2.0=py36hfd86e86_0
-  - lcms2=2.11=h396b838_0
-  - libpng=1.6.37=hbc83047_0
-  - libtiff=4.1.0=h2733197_1
-  - libuuid=1.0.3=h1bed415_2
-  - libxcb=1.14=h7b6447c_0
-  - libxgboost=0.90=he6710b0_1
-  - libxml2=2.9.10=he19cac6_1
-  - lz4-c=1.9.2=he6710b0_1
-  - matplotlib=3.3.1=0
-  - matplotlib-base=3.3.1=py36h817c723_0
-  - networkx=2.5=py_0
-  - olefile=0.46=py_0
-  - openssl=1.1.1h=h7b6447c_0
-  - pcre=8.44=he6710b0_0
-  - pillow=7.2.0=py36hb39fc2d_0
-  - py-xgboost=0.90=py36he6710b0_1
-  - pyparsing=2.4.7=py_0
-  - pyqt=5.9.2=py36h22d08a2_1
-  - python-dateutil=2.8.1=py_0
-  - pywavelets=1.1.1=py36h7b6447c_0
-  - pyyaml=5.3.1=py36h7b6447c_1
-  - qt=5.9.7=h5867ecd_1
-  - scikit-image=0.16.2=py36h0573a6f_0
-  - scikit-learn=0.19.0=py36h97ac459_2
-  - sip=4.19.24=py36he6710b0_0
-  - tk=8.6.10=hbc83047_0
-  - toolz=0.11.1=py_0
-  - tornado=6.0.4=py36h7b6447c_1
-  - xlwt=1.3.0=py36_0
-  - yaml=0.2.5=h7b6447c_0
-  - zstd=1.4.4=h0b5b093_3
-  - progressbar=2.5=py_0
-  - python_abi=3.6=1_cp36m
-  - sk-video=1.1.10=pyh24bf2e0_4
-  - _tflow_select=2.1.0=gpu
-  - absl-py=0.10.0=py36_0
-  - astor=0.8.1=py36_0
-  - c-ares=1.16.1=h7b6447c_0
-  - cairo=1.14.12=h8948797_3
-  - ffmpeg=4.0=hcdf2ecd_0
-  - freeglut=3.0.0=hf484d3e_5
-  - gast=0.4.0=py_0
-  - google-pasta=0.2.0=py_0
-  - graphite2=1.3.14=h23475e2_0
-  - grpcio=1.31.0=py36hf8bcb03_0
-  - h5py=2.8.0=py36h989c5e5_3
-  - harfbuzz=1.8.8=hffaf4a1_0
-  - importlib-metadata=1.7.0=py36_0
-  - jasper=2.0.14=h07fcdf6_1
-  - keras-applications=1.0.8=py_1
-  - keras-preprocessing=1.1.0=py_1
-  - libglu=9.0.0=hf484d3e_1
-  - libopencv=3.4.2=hb342d67_1
-  - libopus=1.3.1=h7b6447c_0
-  - libprotobuf=3.12.4=hd408876_0
-  - libvpx=1.7.0=h439df22_0
-  - markdown=3.2.2=py36_0
-  - ncurses=6.2=he6710b0_1
-  - opencv=3.4.2=py36h6fd60c2_1
-  - pixman=0.40.0=h7b6447c_0
-  - protobuf=3.12.4=py36he6710b0_0
-  - py-opencv=3.4.2=py36hb342d67_1
-  - python=3.6.8=h0371630_0
-  - readline=7.0=h7b6447c_5
-  - tensorboard=1.15.0=pyhb230dea_0
-  - tensorflow-estimator=1.15.1=pyh2649769_0
-  - tensorflow-gpu=1.15.0=h0d30ee6_0
-  - termcolor=1.1.0=py36_1
-  - wrapt=1.12.1=py36h7b6447c_1
-  - zipp=3.1.0=py_0
-  - bzip2=1.0.8=h14c3975_2
-  - hdf5=1.10.2=2
-  - icc_rt=2020.2=intel_254
-  - intel-openmp=2020.2=intel_254
-  - intelpython=2020.2=0
-  - libffi=3.2.1=11
-  - libgcc-ng=9.1.0=hdf63c60_0
-  - libstdcxx-ng=9.1.0=hdf63c60_0
-  - mkl=2020.3=intel_279
-  - mkl-service=2.1.0=py36_2
-  - mkl_fft=1.0.14=py36ha68da19_1
-  - mkl_random=1.0.4=py36ha68da19_2
-  - numpy=1.17.0=py36ha68da19_13
-  - numpy-base=1.17.0=py36_13
-  - pip=19.1.1=py36_0
-  - scipy=1.3.1=py36ha68da19_2
-  - setuptools=41.0.1=py36_0
-  - six=1.12.0=py36_0
-  - sqlite=3.28.0=0
-  - tbb=2020.3=intel_254
-  - tbb4py=2020.3=py36_intel_0
-  - tcl=8.6.4=24
-  - tensorflow=1.15.0=py36_0
-  - tensorflow-base=1.15.2=0
-  - webencodings=0.5.1=py36_0
-  - werkzeug=0.14.1=py36_0
-  - wheel=0.31.0=py36_3
-  - xz=5.2.5=h14c3975_0
-  - zlib=1.2.11.1=h14c3975_1
+  - _libgcc_mutex=0.1
+  - _py-xgboost-mutex=2.0
+  - _tflow_select=2.1.0
+  - absl-py=0.10.0
+  - astor=0.8.1
+  - blas=1.0
+  - bzip2=1.0.8
+  - c-ares=1.16.1
+  - ca-certificates=2020.7.22
+  - cairo=1.14.12
+  - certifi=2020.6.20
+  - cloudpickle=1.6.0
+  - cudatoolkit=10.0.130
+  - cudnn=7.6.5
+  - cupti=10.0.130
+  - cycler=0.10.0
+  - cytoolz=0.11.0
+  - dask-core=2.27.0
+  - dbus=1.13.16
+  - decorator=4.4.2
+  - dill=0.3.2
+  - expat=2.2.9
+  - ffmpeg=4.0
+  - fontconfig=2.13.0
+  - freeglut=3.0.0
+  - freetype=2.10.2
+  - gast=0.2.2
+  - glib=2.65.0
+  - google-pasta=0.2.0
+  - graphite2=1.3.14
+  - grpcio=1.31.0
+  - gst-plugins-base=1.14.0
+  - gstreamer=1.14.0
+  - h5py=2.8.0
+  - harfbuzz=1.8.8
+  - hdf5=1.10.2
+  - icu=58.2
+  - imageio=2.9.0
+  - importlib-metadata=1.7.0
+  - intel-openmp=2020.2
+  - jasper=2.0.14
+  - joblib=0.16.0
+  - jpeg=9b
+  - keras-applications=1.0.8
+  - keras-preprocessing=1.1.0
+  - kiwisolver=1.2.0
+  - lcms2=2.11
+  - ld_impl_linux-64=2.33.1
+  - libedit=3.1.20191231
+  - libffi=3.3
+  - libgcc-ng=9.1.0
+  - libgfortran-ng=7.3.0
+  - libglu=9.0.0
+  - libopencv=3.4.2
+  - libopus=1.3.1
+  - libpng=1.6.37
+  - libprotobuf=3.12.4
+  - libstdcxx-ng=9.1.0
+  - libtiff=4.1.0
+  - libuuid=1.0.3
+  - libvpx=1.7.0
+  - libxcb=1.14
+  - libxgboost=0.90
+  - libxml2=2.9.10
+  - lz4-c=1.9.2
+  - markdown=3.2.2
+  - matplotlib=3.3.1
+  - matplotlib-base=3.3.1
+  - mkl=2020.2
+  - mkl-service=2.3.0
+  - mkl_fft=1.2.0
+  - mkl_random=1.1.1
+  - ncurses=6.2
+  - networkx=2.5
+  - numpy=1.19.1
+  - numpy-base=1.19.1
+  - olefile=0.46
+  - opencv=3.4.2
+  - openssl=1.1.1h
+  - opt_einsum=3.1.0
+  - pcre=8.44
+  - pillow=7.2.0
+  - pip=20.2.2
+  - pixman=0.40.0
+  - progressbar=2.5
+  - protobuf=3.12.4
+  - py-opencv=3.4.2
+  - py-xgboost=0.90
+  - pyparsing=2.4.7
+  - pyqt=5.9.2
+  - python=3.6.12
+  - python-dateutil=2.8.1
+  - python_abi=3.6
+  - pywavelets=1.1.1
+  - pyyaml=5.3.1
+  - qt=5.9.7
+  - readline=8.0
+  - scikit-image=0.16.2
+  - scikit-learn=0.19.0
+  - scipy=1.5.2
+  - setuptools=49.6.0
+  - sip=4.19.24
+  - six=1.15.0
+  - sk-video=1.1.10
+  - sqlite=3.33.0
+  - tensorboard=1.15.0
+  - tensorflow=1.15.0
+  - tensorflow-base=1.15.0
+  - tensorflow-estimator=1.15.1
+  - tensorflow-gpu=1.15.0
+  - termcolor=1.1.0
+  - tk=8.6.10
+  - toolz=0.11.1
+  - tornado=6.0.4
+  - webencodings=0.5.1
+  - werkzeug=0.16.1
+  - wheel=0.35.1
+  - wrapt=1.12.1
+  - xlwt=1.3.0
+  - xz=5.2.5
+  - yaml=0.2.5
+  - zipp=3.1.0
+  - zlib=1.2.11
+  - zstd=1.4.4
   - pip:
     - colour-demosaicing==0.1.5
     - colour-science==0.3.15
     - hmmlearn==0.2.4
-    - scikit-learn==0.23.2
-prefix: /home/djalab/anaconda3/envs/mars
+prefix: /home/mras/miniconda3/envs/mars
+
