+++ MARS_environment_windows.yml
@@ -1,4 +1,4 @@
-name: test
+name: mars
 channels:
   - conda-forge
   - anaconda
@@ -108,6 +108,8 @@ dependencies:
   - tk=8.6.10
   - toolz=0.11.1
   - tornado=6.0.4
+  - typing-extensions=3.7.4.3
+  - typing_extensions=3.7.4.3
   - urllib3=1.25.11
   - vc=14.1
   - vs2015_runtime=14.16.27012
@@ -126,7 +128,8 @@ dependencies:
     - colour-demosaicing==0.1.5
     - colour-science==0.3.15
     - hmmlearn==0.2.4
+    - pyyaml==5.3.1
     - scikit-learn==0.23.2
     - threadpoolctl==2.1.0
-prefix: C:\Users\Ann\miniconda3\envs\mars
+prefix: C:\Users\MARS\miniconda3\envs\mars
